% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_Gibbs}
\alias{C_Gibbs}
\title{A markov sampler using Rcpp}
\usage{
C_Gibbs(n, a, b)
}
\arguments{
\item{n}{Number of random numbers}

\item{a}{Number of random numbers}

\item{b}{Number of random numbers}
}
\value{
Returns random sequence and rejection probability
}
\description{
A random walk Metropolis sampler for generating the standard Laplace distribution
}
\examples{
\dontrun{
C_Gibbs <-(25,2,2)
}
}
