\name{StatComp21030-package}
\alias{StatComp21030-package}
\alias{StatComp21030}
\docType{package}
\title{
  NMF and SNMF
}
\description{
  Realized standard non-negative matrix factorization and sparse non-negative matrix factorization.
}
\details{
  The cbNMF function decomposes a matrix V into the product of two non-negative matrices: V = W * H. The SNMF function further controls the number of non-zero elements in H.
}
\author{
ChengBo.

Maintainer: ChengBo <cb1997@mail.ustc.edu.cn>
}
\keyword{ NMF ,BeSS,SNMF }